processor/requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
